Fixed possable word colision
When making plans for the future, phrases like "We'll have to see what
the weather is like then" might be said. by now checking for "see" the
chance of coliding has been removed .
Removed an old api key for security, and generated a new one (was left
in a comment).